+++ src/main/java/com/fiap/lanchonete/infrastructure/quarkusrest/dto/ExternalInfoPedidoDto.java
@@ -1,5 +1,6 @@
 package com.fiap.lanchonete.infrastructure.quarkusrest.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record ExternalInfoPedidoDto(
@@ -7,6 +8,6 @@ public record ExternalInfoPedidoDto(
         String title,
         String description,
         String notification_url,
-        Double total_amount,
+        BigDecimal total_amount,
         List<ExternalItemsDto> items) {
 }
